:recycle: refactor : 횃불 온도 Double -> Integer 로 변경
900 도가 초기온도이고 높은 만큼 소수점 까지는 필요 없다는 판단 하에 변경

+++ src/main/java/com/c_comachi/inused/domain/post/dto/MainPostInfo.java
@@ -21,7 +21,7 @@ public class MainPostInfo {
     private String tag;
     private String nickname;
     private String profileImage;
-    private Double fireTemperature;
+    private Integer fireTemperature;
 
     private CategoryEntity category;
     private LocalDateTime createdAt;
